# Feat/faster base36 implementation