--- a/go.mod
+++ b/go.mod
@@ -1,8 +1,3 @@
 module github.com/multiformats/go-base36
 
 go 1.11
-
-require (
-	github.com/martinlindhe/base36 v1.0.0
-	github.com/stretchr/testify v1.5.1 // indirect
-)