Ok that will do for now:
```
go-multibase$ go test -count=1 -bench=.
goos: darwin
goarch: amd64
pkg: github.com/multiformats/go-multibase
BenchmarkRoundTrip/base36-8         	  327925	      3683 ns/op
BenchmarkRoundTrip/base36upper-8    	  319947	      3641 ns/op
BenchmarkRoundTrip/base58btc-8      	  386199	      3270 ns/op
BenchmarkRoundTrip/base58flickr-8   	  424201	      2638 ns/op
BenchmarkEncode/base36-8            	  440899	      2785 ns/op
BenchmarkEncode/base36upper-8       	  436546	      2777 ns/op
BenchmarkEncode/base58btc-8         	  755958	      1606 ns/op
BenchmarkEncode/base58flickr-8      	  762080	      1608 ns/op
BenchmarkDecode/base36-8            	 1562305	       769 ns/op
BenchmarkDecode/base36upper-8       	 1558365	       769 ns/op
BenchmarkDecode/base58btc-8         	 1207875	      1057 ns/op
BenchmarkDecode/base58flickr-8      	 1203482	      1041 ns/op
PASS
```